Modification du modele pour long et lat + serializer

+++ pointsEau/serializers.py
@@ -4,6 +4,6 @@ from rest_framework import serializers
 
 class PointEauSerializer(serializers.Serializer):
     nom = serializers.CharField(max_length=100)
-    long = serializers.IntegerField()
-    lat = serializers.IntegerField()
+    long = serializers.DecimalField(max_digits=10, decimal_places=8)
+    lat = serializers.DecimalField(max_digits=10, decimal_places=2)
     desc = serializers.CharField(max_length=255)
